move xrd gui to main larch lib
